Split tests into one test per added option
Added a test for each new option introduced. Also added
default values for the new options added so they don't affect
current formatting without changing these options.

Type: New Feature
