+++ tests/config/indent_switch_pp.cfg
@@ -1,0 +1,7 @@
+pp_indent_case = false
+indent_columns = 4
+indent_with_tabs = 0
+indent_switch_pp = false
+pp_if_indent_code = true
+pp_indent_at_level = true
+pp_indent_count = 4
