--- tests/config/pp_if_indent-5.cfg
@@ -1,15 +1,0 @@
-pp_indent_case       = false
-pp_indent_func_def   = false
-pp_indent_extern     = false
-pp_indent_brace      = false
-pp_indent            = ignore
-indent_columns       = 4
-pp_indent_count      = 4
-pp_if_indent_code    = true
-pp_indent_if         = 0
-pp_indent_at_level   = true
-indent_class         = true
-output_tab_size      = 4
-indent_case_brace    = indent_columns
-indent_switch_pp     = false
-indent_with_tabs     = 0
