+++ tests/config/pp_indent_extern.cfg
@@ -1,0 +1,6 @@
+pp_indent_extern = false
+indent_columns = 4
+indent_with_tabs = 0
+pp_if_indent_code = true
+pp_indent_at_level = true
+pp_indent_count = 4
