+++ tests/input/cpp/pp_indent_brace.cpp
@@ -1,0 +1,25 @@
+//! Create the classiest class
+MyClass::MyClass()
+    : mMember1
+    #if( USE_MEMBER )
+        , mConditionalMember
+    #endif
+{
+    if( isSomething )
+    {
+        DoSomething();
+    }
+
+    #if( USE_FIVE )
+        {
+        DoSomethingAlso();
+        }
+    #endif
+
+    #if( USE_SIX )
+        {
+        Six mySix;
+        DoSomethingWithSix( mySix );
+        }
+    #endif
+}
