+++ tests/input/cpp/pp_indent_extern.cpp
@@ -1,0 +1,11 @@
+// Example for extern "C" block
+#ifdef __cplusplus
+    extern "C" {
+
+    void some_c_function
+    (
+    void
+    );
+
+    }
+#endif
